Updating handleImportError failure case
Removed the NSAssert in handleImportError as we don't
want to fail with an exception when we fail to retrieve
wait time. We would like to keep going. We may choose
later to handle this some other way such as a UI message
but for now the failure is ignored and wait times will
not be shown.