Attempting to add TableView as an alternate view
The tableview is somwhat functional but does have some issues.
If you continue to switch from map to table view, each time
the rows are pushed down more for some reason.

Also, not sure that the way the TableView is added view the xib
is the best way to add it.

Checking into its own branch for safe keeping, sharing, and
refinement.