Fixing TableView content inset bug
Work around problem with translucent toolbar which caused each
display of the tableview to push the bottom content inset up by
the height of the row. Manually setting the bottom inset to
the height of table row seems to fix the issue.

Attempted to find an interface builder option to fix this but
did not see anything that looked fitting.